Extract get_router and add simple test.
This PR extracts the logic from `run_server` that creates a new Router
into a helper function named `get_router` to make it easier to test
without needing to start the full server.

This also adds a simple test that verifies current behavior for
returning a br compressed wasm.wasm file.

This change is in preparation for adding support dynamically change the
available compression.